update DL_logic: create utils.py and move get_image and load model there; simplified predict.py; update requirements; update makefile; update dockerfile for compatibility with silicon chips
