+++ .env.yaml
@@ -1,0 +1,15 @@
+# Model target
+MODEL_TARGET: 'local'
+
+# GCP names
+GCP_PROJECT: 'doggos-101-v2'
+GCP_REGION: 'europe-west1'
+GCR_REGION: 'eu.gcr.io'
+
+# Cloud Storage
+BUCKET_NAME: 'doggos-101'
+
+# API on Artifact Registry
+GAR_IMAGE: 'doggos-101'
+GAR_REPO: 'doggos-101-repo'
+GAR_MEMORY: '4Gi'
