+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM tensorflow/tensorflow:2.10.0
+FROM python:3.8.12-buster
 WORKDIR /prod
 
 # Copy and install requirements
@@ -11,6 +11,4 @@ RUN pip install .
 # Copy the model to container
 COPY models models
 
-COPY container_credential.json container_credential.json
-
-CMD GOOGLE_APPLICATION_CREDENTIALS=container_credential.json uvicorn dogs_prediction.api.fast:app --host 0.0.0.0 --port $PORT
+CMD uvicorn dogs_prediction.api.fast:app --host 0.0.0.0 --port $PORT
