+++ dogs_prediction/api/fast.py
@@ -1,12 +1,11 @@
 from dogs_prediction.DL_logic import predict, registry
 from fastapi import FastAPI, UploadFile
 from fastapi.middleware.cors import CORSMiddleware
-from PIL import Image
 
-#iniate api
+# iniate api
 app = FastAPI()
 
-# Optional, good practice for dev purposes. Allow all middlewares
+# optional, good practice for dev purposes. Allow all middlewares
 app.add_middleware(
     CORSMiddleware,
     allow_origins=['*'],  # Allows all origins
@@ -18,27 +17,23 @@ app.add_middleware(
 # preload the model
 app.state.model = registry.load_convnext_model()
 
-
-
 # predicts from url provided by user
 @app.get('/predict_url')
-def prediction(url_with_pic, model_type='inception_v3'):
+def prediction(url_with_pic):
     model = app.state.model
     assert model is not None
-    prediction = predict.predict_labels(model, model_type, url_with_pic=url_with_pic)
+    prediction = predict.predict_labels(model, url_with_pic=url_with_pic)
     return prediction
-
 
 # predicts from file provided by user
 @app.post('/predict_file')
-def prediction(file: UploadFile, model_type='inception_v3'):
+def prediction(file: UploadFile):
     model = app.state.model
     assert model is not None
-    prediction = predict.predict_labels(model, model_type, img=file.file)
+    prediction = predict.predict_labels(model, img=file.file)
     return prediction
 
-
-#root endpoint
+# root endpoint
 @app.get('/')
 def root():
     return dict(greeting='Bark!')
